+++ src/hooks/useBottomSheet.tsx
@@ -1,10 +1,11 @@
 import { useState, useRef } from 'react';
-import { Animated, Dimensions } from 'react-native';
+import { Animated, Dimensions, PanResponder } from 'react-native';
 
 const useBottomSheet = () => {
   const [isVisible, setIsVisible] = useState(false);
   const screenHeight = Dimensions.get('screen').height;
   const panY = useRef(new Animated.Value(screenHeight)).current;
+
   const translateY = panY.interpolate({
     inputRange: [-1, 0, 1],
     outputRange: [0, 0, 1],
@@ -34,13 +35,29 @@ const useBottomSheet = () => {
     });
   };
 
+  const panResponders = useRef(
+    PanResponder.create({
+      onStartShouldSetPanResponder: () => true,
+      onMoveShouldSetPanResponder: () => false,
+      onPanResponderMove: (event, gestureState) => {
+        panY.setValue(gestureState.dy);
+      },
+      onPanResponderRelease: (event, gestureState) => {
+        if (gestureState.dy > 0 && gestureState.vy > 1) {
+          hideBottomSheet();
+        } else {
+          resetBottomSheet.start();
+        }
+      },
+    }),
+  ).current;
+
   return {
     isVisible,
     showBottomSheet,
     hideBottomSheet,
+    panResponders,
     translateY,
-    resetBottomSheet,
-    panY,
   };
 };
